feat: 위경도 좌표 격자 XY 변환 코드 분리
